refactor: rename @opencode-ai/shared to @opencode-ai/core (#24309)
Pure mechanical rename of the workspace package. packages/shared is
moved to packages/core; all consumers update imports from
@opencode-ai/shared/* to @opencode-ai/core/*. No file content changes
beyond the package rename.

PawWork-local consumers (packages/util, packages/ui, packages/opencode,
packages/desktop-electron runtime-import-guard test) updated.

This is the first half of upstream's "shared → core" refactor; the
follow-up commits that move modules (Global, log, flag, cross-spawn,
npm) into core land in a separate PR (#209 PR2).

Refs: #209
