+++ packages/core/package.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://json.schemastore.org/package.json",
   "version": "1.14.19",
-  "name": "@opencode-ai/shared",
+  "name": "@opencode-ai/core",
   "type": "module",
   "license": "MIT",
   "private": true,
