+++ packages/core/src/global.ts
