+++ packages/core/src/util/array.ts
