+++ packages/core/src/util/binary.ts
