+++ packages/core/src/util/encode.ts
