+++ packages/core/src/util/fn.ts
