+++ packages/core/src/util/hash.ts
