+++ packages/core/src/util/iife.ts
