+++ packages/core/src/util/lazy.ts
