+++ packages/core/src/util/module.ts
