+++ packages/core/test/global.test.ts
