+++ packages/opencode/src/config/config.ts
@@ -19,11 +19,11 @@ import { Event } from "../server/event"
 import { Account } from "@/account"
 import { isRecord } from "@/util/record"
 import type { ConsoleState } from "./console-state"
-import { AppFileSystem } from "@opencode-ai/shared/filesystem"
+import { AppFileSystem } from "@opencode-ai/core/filesystem"
 import { InstanceState } from "@/effect"
 import { makeRuntime } from "@/effect/run-service"
 import { Context, Duration, Effect, Fiber, Layer, Option, Schema } from "effect"
-import { EffectFlock } from "@opencode-ai/shared/util/effect-flock"
+import { EffectFlock } from "@opencode-ai/core/util/effect-flock"
 import { InstanceRef } from "@/effect/instance-ref"
 import { zod, ZodOverride } from "@/util/effect-zod"
 import { ConfigAgent } from "./agent"
@@ -47,7 +47,7 @@ import { Npm } from "@/npm"
 import { Filesystem } from "@/util/filesystem"
 import { Flock } from "@/util/flock"
 import { Installation } from "@/installation"
-import { Runtime } from "@opencode-ai/shared/runtime"
+import { Runtime } from "@opencode-ai/core/runtime"
 
 const log = Log.create({ service: "config" })
 const OPENCODE_PROJECT_CONFIG_NAMES = ["config", "opencode"] as const
