+++ packages/util/test/compat-surface.test.ts
@@ -1,24 +1,24 @@
 import { test, expect } from "bun:test"
 import { findLast as utilFindLast } from "@opencode-ai/util/array"
-import { findLast as sharedFindLast } from "@opencode-ai/shared/util/array"
+import { findLast as sharedFindLast } from "@opencode-ai/core/util/array"
 import { Binary as utilBinary } from "@opencode-ai/util/binary"
-import { Binary as sharedBinary } from "@opencode-ai/shared/util/binary"
+import { Binary as sharedBinary } from "@opencode-ai/core/util/binary"
 import { NamedError as UtilNamedError } from "@opencode-ai/util/error"
-import { NamedError as SharedNamedError } from "@opencode-ai/shared/util/error"
+import { NamedError as SharedNamedError } from "@opencode-ai/core/util/error"
 import { fn as utilFn } from "@opencode-ai/util/fn"
-import { fn as sharedFn } from "@opencode-ai/shared/util/fn"
-import { Identifier as sharedIdentifier } from "@opencode-ai/shared/util/identifier"
+import { fn as sharedFn } from "@opencode-ai/core/util/fn"
+import { Identifier as sharedIdentifier } from "@opencode-ai/core/util/identifier"
 import { Identifier as utilIdentifier } from "@opencode-ai/util/identifier"
 import { iife as utilIife } from "@opencode-ai/util/iife"
-import { iife as sharedIife } from "@opencode-ai/shared/util/iife"
+import { iife as sharedIife } from "@opencode-ai/core/util/iife"
 import { lazy as utilLazy } from "@opencode-ai/util/lazy"
-import { lazy as sharedLazy } from "@opencode-ai/shared/util/lazy"
+import { lazy as sharedLazy } from "@opencode-ai/core/util/lazy"
 import { Module as utilModule } from "@opencode-ai/util/module"
-import { Module as sharedModule } from "@opencode-ai/shared/util/module"
+import { Module as sharedModule } from "@opencode-ai/core/util/module"
 import { retry as utilRetry } from "@opencode-ai/util/retry"
-import { retry as sharedRetry } from "@opencode-ai/shared/util/retry"
+import { retry as sharedRetry } from "@opencode-ai/core/util/retry"
 import { Slug as utilSlug } from "@opencode-ai/util/slug"
-import { Slug as sharedSlug } from "@opencode-ai/shared/util/slug"
+import { Slug as sharedSlug } from "@opencode-ai/core/util/slug"
 import { z } from "zod"
 
 test("shared util surface stays aligned with compatibility util surface", async () => {
