+++ packages/util/test/compat.test.ts
@@ -1,8 +1,8 @@
 import { test, expect } from "bun:test"
 import { base64Encode as utilBase64Encode, checksum as utilChecksum } from "@opencode-ai/util/encode"
-import { base64Encode as sharedBase64Encode, checksum as sharedChecksum } from "@opencode-ai/shared/util/encode"
+import { base64Encode as sharedBase64Encode, checksum as sharedChecksum } from "@opencode-ai/core/util/encode"
 import { getFilename as utilGetFilename } from "@opencode-ai/util/path"
-import { getFilename as sharedGetFilename } from "@opencode-ai/shared/util/path"
+import { getFilename as sharedGetFilename } from "@opencode-ai/core/util/path"
 
 test("util encode and path exports stay compatible with shared", () => {
   const sample = "PawWork 9b"
